- Changed: Some experimental changes to reduce the memory footprint of Game objects.  For this first modification changed regen ticks, race, and speed from int to short.  These few changes reduce the Storm footprint by 2K.  All of these may be changed to ushort but at this time I'm not sure if any of the values can be negative.

+++ GameServer/commands/gmcommands/path.cs
@@ -213,7 +213,7 @@ namespace DOL.GS.Commands
 				DisplayMessage(client, "No path created yet! Use /path create first!");
 				return;
 			}
-			int speed = Math.Min(((GameNPC)client.Player.TargetObject).MaxSpeedBase, path.MaxSpeed);
+			short speed = Math.Min(((GameNPC)client.Player.TargetObject).MaxSpeedBase, (short)path.MaxSpeed);
 
 			// clear any current path
 			((GameNPC)client.Player.TargetObject).CurrentWayPoint = null;
