+++ GameServer/packets/Client/168/UseSkillHandler.cs
@@ -94,11 +94,11 @@ namespace DOL.GS.PacketHandler.Client.v168
 
 				if ((m_flagSpeedData & 0x200) != 0)
 				{
-					player.CurrentSpeed = -(m_flagSpeedData & 0x1ff); // backward movement
+					player.CurrentSpeed = (short)(-(m_flagSpeedData & 0x1ff)); // backward movement
 				}
 				else
 				{
-					player.CurrentSpeed = m_flagSpeedData & 0x1ff; // forwardmovement
+					player.CurrentSpeed = (short)(m_flagSpeedData & 0x1ff); // forwardmovement
 				}
 
 				player.IsStrafing = (m_flagSpeedData & 0x4000) != 0;
